+++ CHANGELOG.md
@@ -39,7 +39,7 @@ Dropping a requirement of a major version of a dependency is a new contract.
 - Deprecate `RelativeTypicalPerformanceJudge` constructor in favor of its `Builder`.
 
 ### Fixed
-- Fix `RelativeNonparametricPerformanceJudge` and `RelativeTypicalPerformanceJudge` treating improvements as noise.
+- Fix `RelativeNonparametricPerformanceJudge` and `RelativeTypicalPerformanceJudge` treating improvements as irrelevant.
 
 [JPERF-1160]: https://ecosystem.atlassian.net/browse/JPERF-1160